norm ?

+++ Libft/ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: piotroff <piotroff@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/03 17:45:27 by apiotrau          #+#    #+#             */
-/*   Updated: 2023/05/11 19:16:10 by piotroff         ###   ########.fr       */
+/*   Updated: 2023/05/15 20:01:44 by piotroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char *ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t i;
 	char *sDup;
 
 	sDup = (char *)s;
